image_processing_old/requirements.txt
filelock==3.13.1
fsspec==2023.12.2
grpcio==1.60.1
grpcio-tools==1.60.1
imageio==2.33.1
imutils==0.5.4
Jinja2==3.1.3
lazy_loader==0.3
MarkupSafe==2.1.4
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.3
opencv-python==4.9.0.80
packaging==23.2
pillow==10.2.0
protobuf==4.25.2
scikit-image==0.22.0
scipy==1.12.0
sympy==1.12
tifffile==2023.12.9
triton==2.1.0
typing_extensions==4.9.0
